Add foreign keys in users and vacations table

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('last_name');
             $table->string('role');
-            $table->integer('department_id')->nullable();
+            $table->unsignedBigInteger('department_id')->nullable();
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
